Rework the time progress elements
