+++ src/lib/module/home_settings/widgets/body/settings_widgets_body.dart
@@ -1,10 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:leafy_launcher/module/home/home_controller.dart';
-import 'package:leafy_launcher/module/home/widget/home_widgets/time_progress/time_progress.dart';
 import 'package:leafy_launcher/module/home_settings/home_settings_controller.dart';
-import 'package:leafy_launcher/module/home_settings/widgets/home_settings_widgets_controller.dart';
-import 'package:leafy_launcher/module/home_settings/widgets/widget/leafy_section_enabled_state_item.dart';
+import 'package:leafy_launcher/module/home_settings/widgets/body/time_progress/time_progress_enabled_state.dart';
+import 'package:leafy_launcher/module/home_settings/widgets/body/time_progress/time_progress_type_state.dart';
 import 'package:leafy_launcher/shared_widget/section/leafy_section_lib.dart';
 import 'package:leafy_launcher/shared_widget/section/src/list/leafy_section_list.dart';
 
@@ -14,9 +12,6 @@ import '../../../../resources/theme/home_theme.dart';
 import '../../../../resources/theme/leafy_theme.dart';
 import '../../../../services/applications/user_applications_controller.dart';
 import '../../../../shared_widget/themed_get_widget.dart';
-
-part 'time_progress/is_on.dart';
-part 'time_progress/type.dart';
 
 class SettingsWidgetsBody
     extends ThemedGetWidget<UserApplicationsController, HomeTheme> {
@@ -30,8 +25,15 @@ class SettingsWidgetsBody
         LeafySection<HomeTheme>(
           header: L10nProvider.getText(L10n.settingsTimeProgressWidget),
           children: const [
-            _IsOn(),
-            _Type(),
+            TimeProgressEnabledState(),
+            TimeProgressTypeState(),
+          ],
+        ),
+        LeafySection<HomeTheme>(
+          header: L10nProvider.getText(L10n.settingsTimeProgressWidget),
+          children: const [
+            TimeProgressEnabledState(),
+            TimeProgressTypeState(),
           ],
         ),
       ],
